Organise classes into multiple packages
