+++ src/main/java/sharva/exceptions/InvalidArgumentsException.java
@@ -1,4 +1,4 @@
-package sharva;
+package sharva.exceptions;
 
 public class InvalidArgumentsException extends SharvaException {
     public InvalidArgumentsException(String message) {
