+++ src/main/java/sharva/exceptions/SharvaException.java
@@ -1,4 +1,4 @@
-package sharva;
+package sharva.exceptions;
 
 public class SharvaException extends Exception {
     public SharvaException(String message) {
