+++ src/main/java/sharva/message/Message.java
@@ -1,4 +1,6 @@
-package sharva;
+package sharva.message;
+
+import sharva.tasks.Task;
 
 public class Message {
     public static final String HORIZONTAL_LINE = "    __________________________________________________________________";
