+++ src/main/java/sharva/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package sharva;
+package sharva.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
